% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_address.R
\name{dist_address}
\alias{dist_address}
\title{Dist_Address}
\usage{
dist_address(locations)
}
\arguments{
\item{locations}{Text addresses to calculate distance between all addresses entered.}
}
\value{
Matrix of distances in miles and returns datafrmae of locations.
}
\description{
Dist_Address
}
