% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meterstomiles.R
\name{meterstomiles}
\alias{meterstomiles}
\title{meterstomiles}
\usage{
meterstomiles(meters)
}
\arguments{
\item{meters}{Single value or vectors of integers.}
}
\value{
Returns the input meters converted to mile units.
}
\description{
Converts Meters to Miles
}
