% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_dbox2.R
\name{nearest_dbox2}
\alias{nearest_dbox2}
\title{Nearest Dbox}
\usage{
nearest_dbox2(lat1d_vec, lon1d_vec, lat2d_vec, lon2d_vec, mode = c("driving",
  "transit", "walking", "bicycling"), value = c("distance_m", "distance_text",
  "duration_s", "duration_text"))
}
\arguments{
\item{lat1d_vec}{Numeric vector, latitude coordinate of voter}

\item{lon1d_vec}{Numeric vector, longitude coordinate of voter}

\item{lat2d_vec}{Numeric vector, latitude coordinate of drop box, polling location}

\item{lon2d_vec}{Numeric vector, longitude coordinate of drop box, polling location}

\item{mode}{Travel mode, one of: "driving" (default), "transit", "walking", "bicycling"}

\item{value}{to extract, one of: "distance_m" (default), "distance_text", "duration_s","duration_text"}
}
\value{
A numeric vector of minimum distances for each voter to their nearest drop box or polling location
}
\description{
Given a set of lat-long coordinates for each voter, and a set of coordinates for all drop boxes or
polling locations, nearest_dbox2() calculates the nearest drop box or polling location for each voter,
in miles or minutes.
}
